Move penalty_coefficient_updaters module inside multiples package
