+++ src/cooper/constraints/__init__.py
@@ -1,3 +1,2 @@
 from .constraint import Constraint
 from .constraint_state import ConstraintState
-from .constraint_type import ConstraintType
