+++ src/cooper/formulations/formulations.py
@@ -5,8 +5,8 @@ import torch
 
 import cooper.formulations.utils as formulation_utils
 from cooper.constraints.constraint_state import ConstraintState
-from cooper.constraints.constraint_type import ConstraintType
 from cooper.multipliers import Multiplier, PenaltyCoefficient
+from cooper.utils import ConstraintType
 
 
 class ContributionStore(NamedTuple):
