--- src/cooper/penalty_coefficient_updaters/__init__.py
@@ -1,1 +1,0 @@
-from .penalty_coefficient_updaters import MultiplicativePenaltyCoefficientUpdater, PenaltyCoefficientUpdater
