+++ src/cooper/utils/__init__.py
@@ -1,2 +1,2 @@
-from .typing import OneOrSequence
+from .typing import ConstraintType, OneOrSequence
 from .utils import ensure_sequence
