+++ tests/multipliers/test_penalty_updater.py
@@ -2,7 +2,7 @@ import pytest
 import torch
 
 import cooper
-from cooper.penalty_coefficient_updaters import MultiplicativePenaltyCoefficientUpdater
+from cooper.multipliers import MultiplicativePenaltyCoefficientUpdater
 
 
 @pytest.fixture(params=[1, 100])
@@ -62,7 +62,7 @@ def test_update_penalty_coefficient_with_unsupported_penalty_coefficient_type(co
     constraint_state = cooper.ConstraintState(violation=torch.tensor(1e-4))
 
     penalty_updater = MultiplicativePenaltyCoefficientUpdater(growth_factor=1.1, violation_tolerance=1e-4)
-    with pytest.raises(ValueError, match=r".*Unsupported penalty coefficient type.*"):
+    with pytest.raises(TypeError, match=r".*Unsupported penalty coefficient type.*"):
         penalty_updater.update_penalty_coefficient_(constraint, constraint_state)
