+++ tutorials/scripts/plot_max_entropy_augmented_lagrangian.py
@@ -28,7 +28,7 @@ import style_utils
 import torch
 
 import cooper
-from cooper.penalty_coefficient_updaters import MultiplicativePenaltyCoefficientUpdater
+from cooper.multipliers import MultiplicativePenaltyCoefficientUpdater
 
 style_utils.set_plot_style()
